+++ internal/forge/forgejo/proposal_merger.go
@@ -1,0 +1,41 @@
+package forgejo
+
+import (
+	"errors"
+	"strconv"
+
+	"codeberg.org/mvdkleijn/forgejo-sdk/forgejo/v2"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	"github.com/git-town/git-town/v21/pkg/colors"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+)
+
+func (self Connector) SquashMergeProposalFn() Option[func(int, gitdomain.CommitMessage) error] {
+	if self.APIToken.IsSome() {
+		return Some(self.squashMergeProposal)
+	}
+	return None[func(int, gitdomain.CommitMessage) error]()
+}
+
+func (self Connector) squashMergeProposal(number int, message gitdomain.CommitMessage) error {
+	if number <= 0 {
+		return errors.New(messages.ProposalNoNumberGiven)
+	}
+	commitMessageParts := message.Parts()
+	self.log.Start(messages.ForgeForgejoMergingViaAPI, colors.BoldGreen().Styled(strconv.Itoa(number)))
+	_, _, err := self.client.MergePullRequest(self.Organization, self.Repository, int64(number), forgejo.MergePullRequestOption{
+		Style:   forgejo.MergeStyleSquash,
+		Title:   commitMessageParts.Subject,
+		Message: commitMessageParts.Text,
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	self.log.Start(messages.APIProposalLookupStart)
+	_, _, err = self.client.GetPullRequest(self.Organization, self.Repository, int64(number))
+	self.log.Ok()
+	return err
+}
