+++ internal/forge/forgejo/proposal_updater.go
@@ -1,0 +1,50 @@
+package forgejo
+
+import (
+	"strconv"
+
+	"codeberg.org/mvdkleijn/forgejo-sdk/forgejo/v2"
+	"github.com/git-town/git-town/v21/internal/forge/forgedomain"
+	"github.com/git-town/git-town/v21/internal/git/gitdomain"
+	"github.com/git-town/git-town/v21/internal/messages"
+	"github.com/git-town/git-town/v21/pkg/colors"
+	. "github.com/git-town/git-town/v21/pkg/prelude"
+)
+
+var _ forgedomain.ProposalUpdater = forgejoConnector
+
+func (self Connector) UpdateProposalBody(proposalData forgedomain.ProposalInterface, newBody string) error {
+	data := proposalData.Data()
+	self.log.Start(messages.APIProposalUpdateBody, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)))
+	_, _, err := self.client.EditPullRequest(self.Organization, self.Repository, int64(data.Number), forgejo.EditPullRequestOption{
+		Body: newBody,
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	return nil
+}
+
+func (self Connector) UpdateProposalTarget(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error {
+	if self.APIToken.IsSome() {
+		return Some(self.updateProposalTarget)
+	}
+	return None[func(forgedomain.ProposalInterface, gitdomain.LocalBranchName) error]()
+}
+
+func (self Connector) updateProposalTarget(proposalData forgedomain.ProposalInterface, target gitdomain.LocalBranchName) error {
+	data := proposalData.Data()
+	targetName := target.String()
+	self.log.Start(messages.APIUpdateProposalTarget, colors.BoldGreen().Styled("#"+strconv.Itoa(data.Number)), colors.BoldCyan().Styled(targetName))
+	_, _, err := self.client.EditPullRequest(self.Organization, self.Repository, int64(data.Number), forgejo.EditPullRequestOption{
+		Base: targetName,
+	})
+	if err != nil {
+		self.log.Failed(err.Error())
+		return err
+	}
+	self.log.Ok()
+	return nil
+}
